requirements_cronjob.txt
langchain==0.0.348
openai==1.3.8
python-dotenv==1.0.0
python-telegram-bot==20.7
youtube-transcript-api==0.6.1
asyncio==3.7.1
